catch improvement

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -15,7 +15,7 @@ class CreateUsersTable extends Migration
     {
         Schema::create('users', function (Blueprint $table) {
             $table->id('id_akun');
-            $table->integer('npwp')->unique();
+            $table->bigInteger('npwp')->unique();
             $table->string('role');
             $table->string('name');
             $table->string('email')->unique();
